Separate issuance from verifier tests.
